style: move correct api to data/local and rename api not repository
