+++ app/src/main/java/kr/co/langquence/data/remote/CorrectApi.kt
@@ -1,10 +1,10 @@
-package kr.co.langquence.model.repository
+package kr.co.langquence.data.remote
 
 import kr.co.langquence.model.domain.CorrectAnswer
 import retrofit2.http.Body
 import retrofit2.http.POST
 
-interface CorrectRepository {
+interface CorrectApi {
 	private companion object {
 		const val CORRECT_API_URL: String = "correct"
 	}
