+++ app/src/main/java/kr/co/langquence/data/repository/CorrectRepositoryImpl.kt
@@ -1,11 +1,11 @@
 package kr.co.langquence.data.repository
 
+import kr.co.langquence.data.remote.CorrectApi
 import kr.co.langquence.model.domain.CorrectAnswer
-import kr.co.langquence.model.repository.CorrectRepository
 import javax.inject.Inject
 
 class CorrectRepositoryImpl @Inject constructor(
-	private val api: CorrectRepository
+	private val api: CorrectApi
 ) {
 	suspend fun request(request: ByteArray): CorrectAnswer = api.request(request)
 }
